Add files via upload
The black and red cables are connected to Ground and +5v. These are connected to -ve and +ve on the relay. The yellow cable is connected to the signal pin on the relay.

When the signal pin is activated (GPIO16), the relay closes the centre silver terminal and the terminal next to the P2 text on the circuit board). When GPIO is off, the centre terminal and the terminal next to the D1 text is closed.